use kwargs and no inheritance for PurchaseView
