+++ src/products/api/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 
+from banking.selector import BANK_CHOICES
 from products.models import Course
 
 
@@ -23,24 +24,19 @@ class CourseSerializer(ShippableSerializer):
         model = Course
 
 
-class PurchaseSerializer(serializers.Serializer):
-    """Simple serializer used to validate purchases"""
-
-    class Meta:
-        fields = [
-            "name",
-            "email",
-        ]
-
-    @classmethod
-    def _validate(cls, input: dict) -> None:
-        instance = cls(data=input)
-        instance.is_valid(raise_exception=True)
-
-
 class CourseSimpleSerializer(serializers.ModelSerializer):
     class Meta:
         model = Course
         fields = [
             "name",
         ]
+
+
+class PurchaseSerializer(serializers.Serializer):
+    name = serializers.CharField(required=True)
+    email = serializers.EmailField(required=True)
+    desired_bank = serializers.ChoiceField(choices=BANK_CHOICES, required=False)
+    promocode = serializers.CharField(max_length=100, required=False)
+    success_url = serializers.CharField(max_length=256, required=False)
+    redirect_url = serializers.CharField(max_length=256, required=False)
+    subscribe = serializers.CharField(max_length=5, required=False)
