+++ src/products/services/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["PurchaseCreator"]
+
+from products.services.purchase_creator import PurchaseCreator
